Add CLI option to choose output directory for vsk run
This commit adds the ability to configure the output SVG directory
when using `vsk run`, via a new command-line option `--output-dir`
(or, equivalently, the `VSK_OUTPUT_DIR`) environment variable.

This provides more flexibility for users using vsketch as a library
with multiple sketches, such as if a user wants to keep saved SVGs
all in a common directory regardless of the directory structure of
the plot source files themselves.